refactor(dashboard): use shared component deploymentID filter
