+++ web/apps/dashboard/lib/schemas/runtime-logs.schema.ts
@@ -15,7 +15,7 @@ export type RuntimeLog = z.infer<typeof dashboardRuntimeLog>;
 export const runtimeLogsRequestSchema = z.object({
   projectId: z.string(),
   appId: z.string().nullable().optional(),
-  deploymentId: z.string().nullable().optional(),
+  deploymentId: z.array(z.string()).optional().default([]),
   environmentId: z
     .object({
       filters: z.array(
