+++ web/apps/dashboard/lib/trpc/routers/deploy/runtime-logs/query.ts
@@ -91,7 +91,7 @@ export const queryRuntimeLogs = workspaceProcedure
       k8sPodNames,
       workspaceId: ctx.workspace.id,
       projectId: project.id,
-      deploymentId: input.deploymentId ?? null,
+      deploymentId: input.deploymentId ?? [],
       appId,
     });
